Hoist the shared firmware-controller stub into tests/conftest.py
